requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.13.0
wandb>=0.15.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
scikit-learn>=1.2.2
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
opencv-python>=4.7.0.72
mtcnn>=0.1.1
pyenchant
nltk
emoji
tqdm
gdown
kaggle
jinja2
python-dotenv
librosa
